made some review changes
