--- packages/ui/src/components/contact/index.ts
@@ -1,4 +1,0 @@
-export { DynamicForm, default } from "./DynamicForm"
-export { default as DynamicFormStep } from "./DynamicFormStep"
-export { fnResolveContactSteps, FIELDS_PER_STEP } from "./contact-form.config"
-export type { TdynamicContactFormProps, TresolvedContactStep } from "./contact-form.types"
